refactor: move subgraph test utilities to a separate module
Extract utility functions from subgraph_validation_tests.rs into subgraph_utils.rs to make them available for other tests in the crate. This allows for more consistent subgraph testing and validation across the codebase.
