+++ apollo-federation/tests/subgraph/mod.rs
@@ -1,2 +1,3 @@
 mod parse_expand_tests;
+pub(crate) mod subgraph_utils;
 mod subgraph_validation_tests;
